Noe mer riktig rating i relevans og lagt til hvor mange engelske artikler fagfeltet er basert på.

+++ App/Models/DomainModels/ResearcherRelevance.cs
@@ -7,6 +7,6 @@ namespace App.Models.DomainModels
 {
     public class ResearcherRelevance : Researcher
     {
-        public int similarities { get; set; }
+        public double similarities { get; set; }
     }
 }
